test(canon): move LSP budget registry check to keep fixtures test under length gate
